Cleanup implementation + add tests
