+++ ddprof-exporter/Cargo.toml
@@ -20,7 +20,7 @@ lazy_static = "1.4"
 libc = "0.2"
 regex = "1.5"
 hyper = { version = "0.14", features = ["http1", "client", "tcp", "stream"], default-features = false }
-tokio = { version = "1.8", features = ["rt"]}
+tokio = { version = "1.8", features = ["rt", "macros"]}
 tokio-rustls = { version = "0.23" }
 percent-encoding = "2.1"
 futures-core = { version = "0.3.0", default-features = false }
