+++ ddprof-exporter/src/errors.rs
@@ -1,7 +1,7 @@
 use std::error;
 use std::fmt;
 
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 #[allow(dead_code)]
 pub(crate) enum Error {
     InvalidUrl,
@@ -17,8 +17,12 @@ impl fmt::Display for Error {
             Self::InvalidUrl => "invalid url",
             Self::OperationTimedOut => "operation timed out",
             Self::UnixSockeUnsuported => "unix sockets unsuported on windows",
-            Self::CannotEstablishTlsConnection => "cannot establish requested secure TLS connection",
-            Self::NoValidCertifacteRootsFound => "native tls couldn't find any valid certifacte roots"
+            Self::CannotEstablishTlsConnection => {
+                "cannot establish requested secure TLS connection"
+            }
+            Self::NoValidCertifacteRootsFound => {
+                "native tls couldn't find any valid certifacte roots"
+            }
         })
     }
 }
